+++ client/src/components/advertisement/advertisementform.scss
@@ -17,6 +17,11 @@
         }
 
         .btn-adv-submit {
+            display: flex;
+            padding: 10px;
+            color: white;
+            justify-content: center;
+            background-color: #065573;
             width: 100%;
         }
     }
