--- client/src/components/leftbar/leftbar.scss
@@ -1,192 +1,0 @@
-@import "../../style.scss";
-
-
- /* I want the leftbar to be display as it is when I'm scrolling */
-    .leftbar {
-        height: 100vh;
-        border-radius: 10px;
-        background: #FFFFFF;
-        // margin-left: 18px;
-        // margin-right: 18px;
-        font-size: 16px;
-        font-weight: bold;
-        justify-content: center;
-        align-items: flex-start;
-       
-        position: sticky;
-
-        .menu {
-            // margin-left: 16px;
-            height: 350px;
-            display: flex;
-            align-items: flex-start;
-            flex-direction: column;
-            justify-content: space-around;
-
-            .menu_item {
-                display: flex;
-                justify-content:flex-start;
-                align-items: center;
-                background-color: #ECFBFE;
-                width: 100%;
-                padding-left: 5px;
-                padding-right: 5px;
-                padding-top: 7px;
-                padding-bottom: 7px;
-                border-radius: 5px;
-
-                img {
-                    padding-left: 10px;
-                    padding-right: 20px;
-                }
-
-                p {
-                    padding-right: 18px;
-                    color: #065573;
-                }
-            }
-
-            .menu_item:hover {
-                background-color: #E5F1F8;
-                cursor: pointer;
-                transform: scale(1.05);
-            }
-        }
-    }
-
-    .hidden_leftbar {
-        width: 100vw;
-        height: max-content;
-        align-items: flex-start;
-        justify-content: space-around;
-        font-size: 16px;
-        font-weight: bold;
-
-        .menu_item_h {
-            background-color: #ECFBFE;
-            border: solid 1px 0px 1px 0px;
-            border-color: black;
-            display: flex;
-            align-items: center;
-            width: 25vw;
-            height: 40px;
-            padding-left: 3px;
-            padding-right: 5px;
-            padding-top: 7px;
-            padding-bottom: 7px;
-
-            img {
-                padding-left: 10px;
-                padding-right: 20px;
-            }
-
-            p {
-                padding-right: 18px;
-            }
-        }
-
-        .menu_item_h:hover {
-            background-color: #D8F6FE;
-            cursor: pointer;
-            transform: scale(1.05);
-        }
-    }
-
-    .hidden_leftbar_s {
-        width: 100vw;
-        height: max-content;
-        align-items: flex-start;
-        justify-content: space-around;
-        font-size: 16px;
-        font-weight: bold;
-
-        .menu_item_h {
-            background-color: #ECFBFE;
-            border: solid 1px 0px 1px 0px;
-            border-color: black;
-            display: flex;
-            align-items: center;
-            width: 25vw;
-            height: 57px;
-            justify-content: space-around;
-
-        }
-
-        .menu_item_h:hover {
-            background-color: #D8F6FE;
-            cursor: pointer;
-            transform: scale(1.05);
-        }
-    }
-
-
-@media only screen and (min-width: 1000px) {
-    .leftbar {
-
-        display: flex;
-    }
-
-    .hidden_leftbar {
-        display: none;
-    }
-
-    .hidden_leftbar_s {
-        display: none;
-    }
-}
-
-@media only screen and (max-width: 1000px) and (min-width: 700px) {
-    .leftbar {
-
-        display: none;
-    }
-
-    .hidden_leftbar {
-        display: flex;
-
-        img {
-            width: 35px;
-        }
-    }
-
-    .hidden_leftbar_s {
-        display: none;
-    }
-}
-
-@media only screen and (max-width: 700px) and (min-width: 500px) {
-    .leftbar {
-        display: none;
-    }
-
-    .hidden_leftbar {
-        display: none;
-    }
-
-    .hidden_leftbar_s {
-        display: flex;
-
-        img {
-            width: 35px;
-        }
-    }
-}
-
-
-@media only screen and (max-width: 500px) {
-    .leftbar {
-        display: none;
-    }
-
-    .hidden_leftbar {
-        display: none;
-    }
-
-    .hidden_leftbar_s {
-        display: flex;
-
-        img {
-            width: 35px;
-        }
-    }
-}
