--- client/src/components/navbar/font.scss
@@ -1,14 +1,0 @@
-/* fonts.css */
-@font-face {
-    font-family: 'Klavika';
-    src: url('../../fonts/klavika-bold.otf') format('opentype');
-    /* Adjust the path based on your project structure */
-    font-weight: normal;
-    font-style: normal;
-  }
-  
-  /* Define a class for the elements where you want to apply the Klavika font */
-  .klavika-h1 {
-    font-family: 'Klavika', sans-serif;
-  }
-
