--- client/src/components/navbar/navbar.scss
@@ -1,147 +1,0 @@
-// klavika
-
-// .navbar{
-//     display: flex;
-//     flex-direction: row;
-//     justify-content: space-between;
-//     align-items: center;
-//     padding: 10px 25px;
-//     background-color: black;
-//     .left{
-//         h1{
-//             color: hwb(0 100% 0%);
-//         }
-//         &.hide {
-//             display: none;
-//           }
-//     }
-//     .right{
-//         .toggle{
-//             border-radius: 3px;
-//             border: 1px solid gray;
-//             cursor: pointer;
-//             span{
-//                 display: block;
-//                 width: 25px;
-//                 height: 2px;
-//                 background-color: white;
-//                 margin: 10px;                
-//             }
-//             &:hover {
-//                 background-color: rgb(234, 234, 234);
-//                 box-shadow: 0 0 4px 1px  rgba(255, 255, 255, 0.7);
-//                 animation: glowing-border 1s infinite;
-//                 span {
-//                   background-color: rgb(0, 0, 0);
-//                 }
-//               }
-//         }
-//     }
-// }
-
-// .navbar{
-//     display: flex;
-//     flex-direction: row;
-//     justify-content: space-between;
-//     padding: 10px 15px;
-//     // box-shadow: 0 4px 10px 2px rgba(0, 0, 0, 0.3);
-//     .left{
-//         img{
-//             margin: 0 10px;
-//         }
-//     }
-//     .right{
-//         img{
-//             margin: 0 10px;
-//         }
-//     }
-//     @media screen and (max-width: 768px) {
-//         .navbar{
-//             .left{
-//                 img{
-//                     width: 45px;
-//                 }
-//             }
-//             .right{
-//                 img{
-//                     width: 45px;
-//                 }
-//             }
-//         }
-//       }
-// }  
-
-.navbar {
-    display: flex;
-    width: 100%;
-    // height: 64px;
-    justify-content: space-between;
-    align-items: center;
-    background: #ffffff;
-
-    .left {
-        display: flex;
-        align-items: center;
-
-        img {
-            padding: 0 10px;
-        }
-    }
-
-    .right {
-        display: flex;
-        align-items: center;
-
-        img {
-            padding: 0 10px;
-        }
-    }
-
-    .popup-container {
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        background: rgba(0, 0, 0, 0.5);
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        z-index: 9999;
-    }
-
-    .popup-content {
-        background: white;
-        padding: 20px;
-        border-radius: 8px;
-        box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-        position: relative;
-
-        button {
-            position: absolute;
-            top: 10px;
-            right: 10px;
-            background: none;
-            border: none;
-            cursor: pointer;
-        }
-    }
-}
-
-
-@media only screen and (max-width: 600px) {
-
-
-    .navbar {
-        img {
-            width: 45px;
-        }
-
-        .left {
-            .logo {
-                width: 20vw;
-            }
-        }
-
-    }
-}
