--- client/src/components/photos/photos.scss
@@ -1,30 +1,0 @@
-.full{
-    display: flex;
-    flex-direction: row;
-    gap: 10px;
-}
-.row{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.box{
-    width: 70px;
-    height: 70px;
-    position: relative;
-    border-radius: 5px;
-    margin: 0 0 5px 5px;
-}
-
-.box img{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 5px;
-}
-
-.box:hover{
-    transform: scale(1.05);
-    cursor: pointer;
-}
