--- client/src/context/authContext.js
@@ -1,59 +1,0 @@
-// import axios from "axios";
-// import { createContext, useEffect, useState } from "react";
-
-// export const AuthContext = createContext();
-
-// export const AuthContextProvider = ({ children }) => {
-//     const [currentUser, setCurrentUser] = useState(
-//         JSON.parse(localStorage.getItem("user")) || null
-//     );
-
-//     const login = async (inputs) => {
-//         const res = await axios.post("http://localhost:5000/server/auth/login", inputs,{
-//             withCredentials: true,
-//         });
-
-//         setCurrentUser(res.data)
-        
-//     };
-
-//     useEffect(() => {
-//         localStorage.setItem("user", JSON.stringify(currentUser));
-//     }, [currentUser]);
-
-//     return (
-//         <AuthContext.Provider value={{ currentUser, login  }}>
-//             {children}
-//         </AuthContext.Provider>
-//     );
-// };
-
-import { createContext, useEffect, useState } from "react";
-
-export const AuthContext = createContext();
-
-export const AuthContextProvider = ({ children }) => {
-    const [currentUser, setCurrentUser] = useState(
-        JSON.parse(localStorage.getItem("user")) || null
-    );
-
-    const login = () => {
-        //TO DO
-        setCurrentUser({
-            id: 1,
-            name: "John Doe",
-            profilePic:
-                "https://images.pexels.com/photos/3228727/pexels-photo-3228727.jpeg?auto=compress&cs=tinysrgb&w=1600",
-        });
-    };
-
-    useEffect(() => {
-        localStorage.setItem("user", JSON.stringify(currentUser));
-    }, [currentUser]);
-
-    return (
-        <AuthContext.Provider value={{ currentUser, login }}>
-            {children}
-        </AuthContext.Provider>
-    );
-};
