+++ client/src/pages/advertisement/advertisementpage.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import "./advertisementpage.scss";
 import Advertisement from "../../components/advertisement/advertisement";
 import "../../components/advertisement/advertisement.scss";
@@ -8,8 +8,18 @@ import hotelpro from "../../assets/images/hotelpro.jpg";
 // import ViewButton from "../../components/advertisement/viewbutton";
 import { IoOptions } from "react-icons/io5";
 import LeftBar from "../../components/leftbar/leftbar";
+import Filter from "../../components/filter/filter";
 
 export const AdvertisementPage = () => {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const handleButtonClick = () => {
+    setIsModalOpen(true);
+  };
+
+  const closeModal = () => {
+    setIsModalOpen(false);
+  };
   return (
     <div className="hotel-profile">
       <div className="left">
@@ -28,7 +38,17 @@ export const AdvertisementPage = () => {
         </div>
         <div className="add-ad">
           <div className="filter">
-            <IoOptions title="Filter" />
+            <IoOptions title="Filter" onClick={handleButtonClick} />
+            {isModalOpen && (
+        <div className="modal">
+          <div className="modal-content">
+            <div className="close-button" onClick={closeModal}>
+              <span>&times;</span>
+            </div>
+            <Filter />
+          </div>
+        </div>
+      )}
           </div>
           <Addbutton />
         </div>
