--- client/src/pages/home/home.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const Home = () => {
-  return (
-    <div>Home</div>
-  )
-}
-
-export default Home;
