--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "roamly",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mui/icons-material": "^5.14.0",
-    "@mui/material": "^5.14.0",
-    "bootstrap": "^5.3.0",
-    "react-bootstrap": "^2.8.0",
-    "react-router-dom": "^6.14.1",
-    "sass": "^1.63.6"
-  }
-}
