--- server/controllers/user.js
@@ -1,3 +1,0 @@
-export const getUser = (req, res)=>{
-
-}
