--- server/routers/auth.js
@@ -1,10 +1,0 @@
-import express from "express";
-import { login, signup, logout } from "../controllers/auth.js";
-
-const router = express.Router();
-
-router.post("/login", login);
-router.post("/signup", signup);
-router.post("/logout", logout);
-
-export default router;
