--- server/routers/users.js
@@ -1,8 +1,0 @@
-import express from "express";
-import { getUser } from "../controllers/user.js";
-
-const router = express.Router();
-
-router.get("/find/:userId", getUser)
-
-export default router;
